+++ components/admin/Home/index.tsx
@@ -1,0 +1,9 @@
+import React from 'react';
+
+interface Props {}
+
+const AdminHome = ({}: Props) => {
+  return <h1>Admin home</h1>;
+};
+
+export default AdminHome;