+++ components/admin/User/index.tsx
@@ -1,0 +1,9 @@
+import React from 'react';
+
+interface Props {}
+
+const AdminAccount = ({}: Props) => {
+  return <h1>Admin Account</h1>;
+};
+
+export default AdminAccount;