+++ components/util/IsLoggedIn.tsx
@@ -4,6 +4,7 @@ import {UserState} from '../../store/user/reducers';
 
 interface Props {
   children: ReactNode;
+  noneLoggedInView?: any;
   userSession: UserState;
 }
 
@@ -11,10 +12,22 @@ const isUserLoggedIn = (userSession: UserState) => {
   return userSession.token.length > 0;
 };
 
-const IsLoggedIn = ({children, userSession, ...props}: Props) => {
+const IsLoggedIn = ({
+  children,
+  noneLoggedInView: Element = null,
+  userSession,
+  ...props
+}: Props) => {
   const loggedIn = isUserLoggedIn(userSession);
   if (loggedIn) {
     return <React.Fragment>{children}</React.Fragment>;
+  }
+  if (Element) {
+    return (
+      <React.Fragment>
+        <Element />
+      </React.Fragment>
+    );
   }
   return <React.Fragment />;
 };