+++ data/consts.ts
@@ -2,6 +2,20 @@ export const LOGIN_EXPIRY_SECONDS = 86400;
 
 export const DEFAUlT_IMG = 'default.png';
 
+export const ROUTES = {
+  ADMIN: '/admin?section=home',
+  HOME: '/',
+  LOGIN: '/login',
+};
+
+export const ADMIN_SECTIONS = {
+  HOME: 'home',
+  BLOG_HOME: 'blog',
+  BLOG_ADD: 'blog-add',
+  BLOG_EDIT: 'blog-edit',
+  USER: 'user',
+};
+
 export const PAGE_LAYERS = {
   TOP_VIEW: 5000,
   OVERLAY: 4000,