+++ CHANGELOG.md
@@ -4,6 +4,10 @@ Changes will be documented here generally around release items using symantic ve
 
 ## Unreleased
 
+- [login-page] working on designing the login page
+- [admin] building out the admin section of the site
+  - create subdomains
+
 ## Version 1.0.0
 
 ## Version 1.0.0-rc